// src/shared/api/schema.ts
export interface components {
    schemas: {
// ...
        UpdatePlaylistRequestPayload: {
            data: components["schemas"]["UpdatePlaylistData"];
        };
// ...
        GetPlaylistsOutput: {
            /** @description Array of playlist resource objects */
            data: components["schemas"]["PlaylistListItemResource"][];
            meta: components["schemas"]["JsonApiMetaWithPaging"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}